Move js out from assets folder
