+++ src/components/Toolbar.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import TypeOfIssue from "../assets/js/type-of-issue";
+import TypeOfIssue from "../enums/type-of-issue";
 
 export default {
   props: {
