+++ src/handlers/sprint-search-handler.js
@@ -1,7 +1,7 @@
-import Jira from '../../configs/vue-resource';
+import Jira from '../configs/vue-resource';
 import IssueHandler from './issue-handler'
-import TypeOfIssue from './type-of-issue'
-import Optional from '../../helpers/optional'
+import TypeOfIssue from '../enums/type-of-issue'
+import Optional from '../helpers/optional'
 
 export default class SprintSearchHandler {
   constructor(sprintName) {
